pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytradingview"
version = "0.1.0"
description = "A Python client for TradingView WebSocket API. It's a lightweight library capable of providing real-time data on stocks, cryptocurrencies, indices etc."
authors = [{ name = "Onesimus Graves-Sampson", email = "onesimusgravessampson1@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["websocket-client>=1.5.1"]
